+++ blueprints/flexberry-list-form/index.js
@@ -7,8 +7,7 @@ var path = require('path');
 var lodash = require('lodash');
 var stripBom = require("strip-bom");
 var Locales_1 = require('../flexberry-core/Locales');
-var Blueprint = require('ember-cli/lib/models/blueprint');
-var AddonBlueprint_1 = require('../flexberry-addon/AddonBlueprint');
+var CommonUtils_1 = require('../flexberry-common/CommonUtils');
 module.exports = {
     description: 'Generates an ember list-form for flexberry.',
     availableOptions: [
@@ -23,21 +22,17 @@ module.exports = {
         if (this._files) {
             return this._files;
         }
-        this._super._files = null;
-        this._super.path = this.path;
-        this._files = this._super.files();
-        this._super._files = null;
         if (this.options.dummy) {
-            lodash.remove(this._files, function (v) { return v === "app/templates/__name__.hbs" || v === "app/templates/__name__/loading.hbs"; });
+            this._files = CommonUtils_1.default.getFilesForGeneration(this, function (v) { return v === "app/templates/__name__.hbs" || v === "app/templates/__name__/loading.hbs"; });
         }
         else {
-            lodash.remove(this._files, function (v) { return v === "tests/dummy/app/templates/__name__.hbs" || v === "tests/dummy/app/templates/__name__/loading.hbs"; });
+            this._files = CommonUtils_1.default.getFilesForGeneration(this, function (v) { return v === "tests/dummy/app/templates/__name__.hbs" || v === "tests/dummy/app/templates/__name__/loading.hbs"; });
         }
         return this._files;
     },
     afterInstall: function (options) {
         if (this.project.isEmberCLIAddon()) {
-            AddonBlueprint_1.default.install(options, ["controller", "route"]);
+            CommonUtils_1.default.installFlexberryAddon(options, ["controller", "route"]);
         }
     },
     /**