+++ blueprints/flexberry-object/files/__root__/mixins/regenerated/objects/__name__.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
 
 export let ObjectMix = Ember.Mixin.create({
-  <%= attrs %>
+<%= attrs %>
 });