+++ blueprints/flexberry-serializer-init/files/__root__/serializers/__name__.js
@@ -1,4 +1,5 @@
-import { Serializer as <%= className %>Serializer } from '../mixins/regenerated/serializers/<%= name %>';
+import { Serializer as <%= className %>Serializer } from
+  '../mixins/regenerated/serializers/<%= name %>';
 import <%if(parentModelName) {%><%=parentClassName%>Serializer from './<%=parentModelName%>';<%}else{%>__ApplicationSerializer from './application';<%}%>
 
 export default <%if(parentModelName) {%><%=parentClassName%>Serializer<%}else{%>__ApplicationSerializer<%}%>.extend(<%= className %>Serializer, {